+++ src/function/arithmetic/largereq.js
@@ -1,29 +1,29 @@
 /**
- * Check if value x is larger y, x > y
+ * Check if value x is larger or equal to y, x >= y
  * In case of complex numbers, the absolute values of a and b are compared.
  * @param  {Number | Complex | Unit | String} x
  * @param  {Number | Complex | Unit | String} y
  * @return {Boolean} res
  */
-function larger(x, y) {
+function largereq(x, y) {
     if (arguments.length != 2) {
-        throw newArgumentsError('larger', arguments.length, 2);
+        throw newArgumentsError('largereq', arguments.length, 2);
     }
 
     if (isNumber(x)) {
         if (isNumber(y)) {
-            return x > y;
+            return x >= y;
         }
         else if (y instanceof Complex) {
-            return x > abs(y);
+            return x >= abs(y);
         }
     }
     if (x instanceof Complex) {
         if (isNumber(y)) {
-            return abs(x) > y;
+            return abs(x) >= y;
         }
         else if (y instanceof Complex) {
-            return abs(x) > abs(y);
+            return abs(x) >= abs(y);
         }
     }
 
@@ -31,44 +31,42 @@ function larger(x, y) {
         if (!x.equalBase(y)) {
             throw new Error('Cannot compare units with different base');
         }
-        return x.value > y.value;
+        return x.value >= y.value;
     }
 
     if (isString(x) || isString(y)) {
-        return x > y;
+        return x >= y;
     }
 
     // TODO: implement array support
     // TODO: implement matrix support
 
-    throw newUnsupportedTypeError('larger', x, y);
+    throw newUnsupportedTypeError('largereq', x, y);
 }
 
-math.larger = larger;
+math.largereq = largereq;
 
 /**
  * Function documentation
  */
-larger.doc = {
-    'name': 'larger',
+largereq.doc = {
+    'name': 'largereq',
     'category': 'Operators',
     'syntax': [
-        'x > y',
-        'larger(x, y)'
+        'x >= y',
+        'largereq(x, y)'
     ],
     'description':
-        'Check if value x is larger than y. ' +
-        'Returns 1 if x is larger than y, and 0 if not.',
+        'Check if value x is larger or equal to y. ' +
+        'Returns 1 if x is larger or equal to y, and 0 if not.',
     'examples': [
-        '2 > 3',
-        '5 > 2*2',
-        'a = 3.3',
+        '2 > 1+1',
+        '2 >= 1+1',
+        'a = 3.2',
         'b = 6-2.8',
-        '(a > b)',
-        '(b < a)',
-        '5 cm > 2 inch'
+        '(a > b)'
     ],
     'seealso': [
-        'equal', 'unequal', 'smaller', 'smallereq', 'largereq'
+        'equal', 'unequal', 'smallereq', 'smaller', 'largereq'
     ]
 };