+++ CHANGELOG.md
@@ -5,7 +5,8 @@ https://github.com/josdejong/mathjs
 ## <not yet released>, version 0.2.0
 
 - Parser, Scope, and Node tree implemented.
-- Implemented methods import, re, im, arg, conj.
+- Implemented methods import, re, im, arg, conj, mod, equal, unequal,
+  smallereq, largereq.
 
 
 ## 2013-02-18, version 0.1.0